Fix David's mess - Cleaned up UI a lot - Fixed some style violations - Moved contact_model to database - Added floating action buttons - Fixed storage permission on android 10 (also see #169) - Added TODOs for more things to fix later
