+++ lib/common/avatars.dart
@@ -5,9 +5,9 @@ import 'package:multacc/common/theme.dart';
 
 /// Contains static methods that return proper CircleAvatar widgets
 class Avatars {
-  static Widget buildContactAvatar({@required Uint8List memoryImage, double radius = 20.0}) {
+  static Widget buildContactAvatar({@required Uint8List memoryImage, IconData defaultIcon = Icons.person, double radius = 20.0}) {
     return memoryImage != null && memoryImage.length > 0
         ? CircleAvatar(backgroundImage: MemoryImage(memoryImage), radius: radius, backgroundColor: Colors.transparent)
-        : CircleAvatar(child: Icon(Icons.person, size: radius), backgroundColor: kBackgroundColorLight, radius: radius);
+        : CircleAvatar(child: Icon(defaultIcon, size: radius), backgroundColor: kBackgroundColorLight, radius: radius);
   }
 }
