+++ lib/common/search_delegate.dart
@@ -4,7 +4,7 @@ import 'package:get_it/get_it.dart';
 import 'package:multacc/common/avatars.dart';
 import 'package:multacc/common/theme.dart';
 import 'package:multacc/pages/contacts/contact_details_page.dart';
-import 'package:multacc/pages/contacts/contact_model.dart';
+import 'package:multacc/database/contact_model.dart';
 import 'package:multacc/pages/contacts/contacts_data.dart';
 
 /// Helper class for searching contacts
