+++ lib/database/database_interface.dart
@@ -2,7 +2,7 @@ import 'package:hive/hive.dart';
 import 'package:hive_flutter/hive_flutter.dart';
 
 import 'package:multacc/items/item.dart';
-import 'package:multacc/pages/contacts/contact_model.dart';
+import 'package:multacc/database/contact_model.dart';
 
 class DatabaseInterface {
   Box<MultaccContact> get contactsBox => Hive.box<MultaccContact>('contacts');
