+++ lib/main.dart
@@ -14,8 +14,6 @@ import 'package:multacc/database/database_interface.dart';
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
 
-  // final dbName = "Multacc_Database.db";
-
   // request contacts permission
   while (!await Permission.contacts.request().isGranted) {}
