+++ lib/pages/contacts/contact_details_page.dart
@@ -4,8 +4,7 @@ import 'package:multacc/common/avatars.dart';
 import 'package:multacc/common/theme.dart';
 import 'package:multacc/items/item.dart';
 import 'package:multacc/items/phone.dart';
-import 'package:multacc/pages/contacts/contact_form.dart';
-import 'package:multacc/pages/contacts/contact_model.dart';
+import 'package:multacc/database/contact_model.dart';
 
 class ContactDetailsPage extends StatefulWidget {
   final MultaccContact contact;
@@ -27,43 +26,16 @@ class _ContactDetailsPageState extends State<ContactDetailsPage> {
 
     return Material(
       child: Container(
-        height: MediaQuery.of(context).size.height / 1.1, // @todo Fix ContactDetails height for scrolling
+        height: MediaQuery.of(context).size.height / 1.2, // @todo Fix ContactDetails height for scrolling
         child: Column(
           children: <Widget>[
-            _buildHeader(),
+            Avatars.buildContactAvatar(memoryImage: contact.avatar, radius: 40.0),
             _buildName(),
-            _buildShortcutsRow(),
+            // _buildShortcutsRow(),
             _buildContactItemsList(),
           ],
         ),
       ),
-    );
-  }
-
-  Widget _buildHeader() {
-    return Row(
-      mainAxisAlignment: MainAxisAlignment.center,
-      children: <Widget>[
-        Container(width: MediaQuery.of(context).size.width / 3),
-        Container(
-          width: MediaQuery.of(context).size.width / 3,
-          child: Avatars.buildContactAvatar(memoryImage: contact.avatar, radius: 40.0),
-        ),
-        Container(
-          width: MediaQuery.of(context).size.width / 3,
-          child: IconButton(
-            icon: Icon(Icons.edit, color: Colors.grey),
-            onPressed: () {
-              Navigator.of(context).push(MaterialPageRoute<Null>(
-                builder: (BuildContext context) {
-                  return ContactForm(contact);
-                },
-                fullscreenDialog: true
-              ));
-            },
-          ),
-        ),
-      ],
     );
   }
 
@@ -88,7 +60,8 @@ class _ContactDetailsPageState extends State<ContactDetailsPage> {
     );
   }
 
-  Padding _buildShortcutsRow() {
+  // @todo Fix the terrible contact shortcuts row
+  Widget _buildShortcutsRow() {
     return Padding(
       padding: const EdgeInsets.symmetric(vertical: 12.0),
       child: Row(
